fix(snapvault): correct photoElement reference in loadOriginalImage
TypeError: Cannot read properties of undefined (reading 'photoElement')
at Lightbox.loadOriginalImage (lightbox.js:504:37)

The method was incorrectly accessing this.lightbox.photoElement when
this is already the Lightbox instance. Changed to this.photoElement.

Line 504:
  - Before: const photo = this.lightbox.photoElement;
  - After:  const photo = this.photoElement;

Context: In loadOriginalImage(), 'this' refers to the Lightbox instance
(see constructor line 14-50 where photoElement is defined as this.photoElement)